backend/pyproject.toml
[tool.poetry]
name = "anirecs"
version = "0.1.0"
description = ""
authors = ["ejedavid <ejeohejidavid@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.110.2"
pydantic = "^2.3.0"
python-jose = "^3.3.0" 
sqlalchemy = "^1.4.0"
uvicorn = "^0.29.0"
pydantic-settings = "^2.2.1"
psycopg2-binary = "^2.9.9"
passlib = "^1.7.4"
httpx = "^0.27.0"
mutmut = "^2.4.5"
python-dotenv = "^1.0.1"
pytest = "^8.2.0"
pytest-cov = "^5.0.0"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.0.0"
bandit = "^1.7.8"
pytest = "^8.2.0"
pytest-cov = "^5.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
